Fix time travel debugging in vuex
Use vuex-router-sync to keep route state in store to be able to use time
travel debugging with a working routing. Without this package the time
travel worked except that the url in the browser was never updated.

It also enables us to reuse the query string parameter from the route
object in the store when updating the model by content link.

Feature: CMS-11254