Apply application path on request from client
When an app is configured in iis to be hosted as a virtual directory we
need to tell the client about the application path so that it can create
client side requests with correct URLs. This is used when updating the
model after editing it when it's hosted as a virtual directory.

Fixes: CMS-12515