Use the new data-epi-edit attribute
This attribute was created to make it easier to get On-Page Editing for
client-side rendered websites. Setting `data-epi-edit="PropertyName"` is
equivalent to setting these three properties:
* data-epi-property-name="PropertyName"
* data-epi-property-render="none"
* data-epi-property-edittype="floating"

All three are otherwise needed when doing client-side rendering so the CMS
doesn't update the DOM and break its connection to the client-side rendering
framework.

Task: CMS-12661
Feature: CMS-12641