Make blocks editable in ContentAreas on OPE
Set the block id with "data-epi-block-id" to get the overlays attached.
The id is available on the model in the store (which is given by the
ContentDeliveryAPI).

Only set "data-epi-block-id" when in OPE (basically when the context
isEditable). This is to avoid leaking too much info when in View Mode
(what the visitor sees).

Task: CMS-12892
Feature: CMS-12819